Issue #100: implemented and tested.
