+++ pixels-core/src/main/java/io/pixelsdb/pixels/core/writer/ColumnWriter.java
@@ -20,6 +20,7 @@
 package io.pixelsdb.pixels.core.writer;
 
 import io.pixelsdb.pixels.core.PixelsProto;
+import io.pixelsdb.pixels.core.TypeDescription;
 import io.pixelsdb.pixels.core.stats.StatsRecorder;
 import io.pixelsdb.pixels.core.vector.ColumnVector;
 
@@ -32,6 +33,48 @@ import java.io.IOException;
  */
 public interface ColumnWriter
 {
+    /**
+     * Create a column writer according to the data type.
+     * @param schema the data type.
+     * @param pixelStride
+     * @param isEncoding set true if enable data encoding.
+     * @return
+     */
+    public static ColumnWriter newColumnWriter(TypeDescription schema, int pixelStride, boolean isEncoding)
+    {
+        switch (schema.getCategory())
+        {
+            case BOOLEAN:
+                return new BooleanColumnWriter(schema, pixelStride, isEncoding);
+            case BYTE:
+                return new ByteColumnWriter(schema, pixelStride, isEncoding);
+            case SHORT:
+            case INT:
+            case LONG:
+                return new IntegerColumnWriter(schema, pixelStride, isEncoding);
+            case FLOAT:
+                return new FloatColumnWriter(schema, pixelStride, isEncoding);
+            case DOUBLE:
+                return new DoubleColumnWriter(schema, pixelStride, isEncoding);
+            case STRING:
+                return new StringColumnWriter(schema, pixelStride, isEncoding);
+            case CHAR:
+                return new CharColumnWriter(schema, pixelStride, isEncoding, schema.getMaxLength());
+            case VARCHAR:
+                return new VarcharColumnWriter(schema, pixelStride, isEncoding, schema.getMaxLength());
+            case BINARY:
+                return new BinaryColumnWriter(schema, pixelStride, isEncoding);
+            case DATE:
+                return new DateColumnWriter(schema, pixelStride, isEncoding);
+            case TIME:
+                return new TimeColumnWriter(schema, pixelStride, isEncoding);
+            case TIMESTAMP:
+                return new TimestampColumnWriter(schema, pixelStride, isEncoding);
+            default:
+                throw new IllegalArgumentException("Bad schema type: " + schema.getCategory());
+        }
+    }
+
     int write(ColumnVector vector, int length)
             throws IOException;
